split prompt wallet into multiple files
