+++ neo/Prompt/Commands/Wallet.py
@@ -1,7 +1,5 @@
 from neo.Core.Blockchain import Blockchain
-from neo.Wallets import NEP5Token
 from neo.Core.TX.ClaimTransaction import ClaimTransaction
-from neo.Core.TX.Transaction import ContractTransaction
 from neo.Core.TX.Transaction import TransactionOutput
 from neo.Core.TX.TransactionAttribute import TransactionAttribute, TransactionAttributeUsage
 from neo.SmartContract.ContractParameterContext import ContractParametersContext
@@ -11,35 +9,20 @@ from neo.Wallets.utils import to_aes_key
 from neo.Implementations.Wallets.peewee.UserWallet import UserWallet
 from neocore.Fixed8 import Fixed8
 from neocore.UInt160 import UInt160
-from neo.SmartContract.Contract import Contract
-from neocore.Cryptography.Crypto import Crypto
-from neocore.KeyPair import KeyPair
 from prompt_toolkit import prompt
-import binascii
 import json
 import os
-from neo.Implementations.Wallets.peewee.Models import Account
 from neo.Prompt.CommandBase import CommandBase, CommandDesc, ParameterDesc
 from neo.Prompt.PromptData import PromptData
 from neo.Prompt.Commands.Send import CommandWalletSend, CommandWalletSendMany, CommandWalletSign
 from neo.Prompt.Commands.Tokens import CommandWalletToken
-from neo.Prompt.Commands.LoadSmartContract import ImportContractAddr
+from neo.Prompt.Commands.WalletAddress import CommandWalletAddress
+from neo.Prompt.Commands.WalletImport import CommandWalletImport
+from neo.Prompt.Commands.WalletExport import CommandWalletExport
 from neo.logging import log_manager
 from neocore.Utils import isValidPublicAddress
 
 logger = log_manager.getLogger()
-
-
-def _is_valid_public_key(key):
-    if len(key) != 66:
-        return False
-    try:
-        Crypto.ToScriptHash(key, unhex=True)
-    except Exception:
-        # the UINT160 inside ToScriptHash can throw Exception
-        return False
-    else:
-        return True
 
 
 class CommandWallet(CommandBase):
@@ -50,19 +33,16 @@ class CommandWallet(CommandBase):
         self.register_sub_command(CommandWalletOpen())
         self.register_sub_command(CommandWalletClose())
         self.register_sub_command(CommandWalletVerbose(), ['v', '--v'])
-        self.register_sub_command(CommandWalletCreateAddress())
-        self.register_sub_command(CommandWalletDeleteAddress())
         self.register_sub_command(CommandWalletSend())
         self.register_sub_command(CommandWalletSendMany())
         self.register_sub_command(CommandWalletSign())
         self.register_sub_command(CommandWalletClaimGas())
         self.register_sub_command(CommandWalletRebuild())
         self.register_sub_command(CommandWalletUnspent())
-        self.register_sub_command(CommandWalletSplit())
-        self.register_sub_command(CommandWalletAlias())
         self.register_sub_command(CommandWalletToken())
         self.register_sub_command(CommandWalletExport())
         self.register_sub_command(CommandWalletImport())
+        self.register_sub_command(CommandWalletAddress())
 
     def command_desc(self):
         return CommandDesc('wallet', 'manage wallets')
@@ -203,43 +183,6 @@ class CommandWalletVerbose(CommandBase):
         return CommandDesc('verbose', 'show additional wallet details')
 
 
-class CommandWalletCreateAddress(CommandBase):
-
-    def __init__(self):
-        super().__init__()
-
-    def execute(self, arguments):
-        addresses_to_create = PromptUtils.get_arg(arguments, 0)
-
-        if not addresses_to_create:
-            print("Please specify a number of addresses to create.")
-            return
-
-        return CreateAddress(PromptData.Wallet, addresses_to_create)
-
-    def command_desc(self):
-        p1 = ParameterDesc('number of addresses', 'number of addresses to create')
-        return CommandDesc('create_addr', 'create a new wallet address', params=[p1])
-
-
-class CommandWalletDeleteAddress(CommandBase):
-    def __init__(self):
-        super().__init__()
-
-    def execute(self, arguments):
-        addr_to_delete = PromptUtils.get_arg(arguments, 0)
-
-        if not addr_to_delete:
-            print("Please specify an address to delete.")
-            return False
-
-        return DeleteAddress(PromptData.Wallet, addr_to_delete)
-
-    def command_desc(self):
-        p1 = ParameterDesc('address', 'address to delete')
-        return CommandDesc('delete_addr', 'delete a wallet address', params=[p1])
-
-
 class CommandWalletClaimGas(CommandBase):
 
     def __init__(self):
@@ -319,515 +262,8 @@ class CommandWalletUnspent(CommandBase):
         return CommandDesc('unspent', 'show unspent assets', params=[p1, p2, p3, p4])
 
 
-class CommandWalletSplit(CommandBase):
-
-    def __init__(self):
-        super().__init__()
-
-    def execute(self, arguments):
-        wallet = PromptData.Wallet
-
-        if len(arguments) < 4:
-            print("Please specify the required parameters")
-            return
-
-        if len(arguments) > 5:
-            # the 5th argument is the optional attributes,
-            print("Too many parameters supplied. Please check your command")
-            return
-
-        addr = arguments[0]
-        if not isValidPublicAddress(addr):
-            print("Invalid address specified")
-            return
-
-        try:
-            from_addr = wallet.ToScriptHash(addr)
-        except ValueError as e:
-            print(str(e))
-            return
-
-        asset_id = PromptUtils.get_asset_id(wallet, arguments[1])
-        if not asset_id:
-            print(f"Unknown asset id: {arguments[1]}")
-            return
-
-        try:
-            index = int(arguments[2])
-        except ValueError:
-            print(f"Invalid unspent index value: {arguments[2]}")
-            return
-
-        try:
-            divisions = int(arguments[3])
-        except ValueError:
-            print(f"Invalid divisions value: {arguments[3]}")
-            return
-
-        if divisions < 2:
-            print("Divisions cannot be lower than 2")
-            return
-
-        if len(arguments) == 5:
-            fee = Fixed8.TryParse(arguments[4], require_positive=True)
-            if not fee:
-                print(f"Invalid fee value: {arguments[4]}")
-                return
-        else:
-            fee = Fixed8.Zero()
-
-        return SplitUnspentCoin(wallet, asset_id, from_addr, index, divisions, fee)
-
-    def command_desc(self):
-        p1 = ParameterDesc('address', 'address to split from')
-        p2 = ParameterDesc('asset', 'type of asset to split (NEO/GAS)')
-        p3 = ParameterDesc('unspent_index', 'index of the vin to split')
-        p4 = ParameterDesc('divisions', 'divide into number of vouts')
-        p5 = ParameterDesc('fee', 'optional fee', optional=True)
-        return CommandDesc('split', 'split an asset unspent output into N outputs', params=[p1, p2, p3, p4, p5])
-
-
-class CommandWalletAlias(CommandBase):
-
-    def __init__(self):
-        super().__init__()
-
-    def execute(self, arguments):
-        if len(arguments) < 2:
-            print("Please supply an address and an alias")
-            return False
-
-        return AddAlias(PromptData.Wallet, arguments[0], arguments[1])
-
-    def command_desc(self):
-        p1 = ParameterDesc('address', 'address to create an alias for')
-        p2 = ParameterDesc('alias', 'alias to associate with the address')
-        return CommandDesc('alias', 'create an alias for an address', params=[p1, p2])
-
-
-class CommandWalletExport(CommandBase):
-
-    def __init__(self):
-        super().__init__()
-        self.register_sub_command(CommandWalletExportWIF())
-        self.register_sub_command(CommandWalletExportNEP2())
-
-    def command_desc(self):
-        return CommandDesc('export', 'export wallet items')
-
-    def execute(self, arguments):
-        item = PromptUtils.get_arg(arguments)
-
-        if not item:
-            print(f"Please specify an action. See help for available actions")
-            return False
-
-        try:
-            return self.execute_sub_command(item, arguments[1:])
-        except KeyError:
-            print(f"{item} is an invalid parameter")
-        return False
-
-
-class CommandWalletImport(CommandBase):
-
-    def __init__(self):
-        super().__init__()
-        self.register_sub_command(CommandWalletImportWIF())
-        self.register_sub_command(CommandWalletImportNEP2())
-        self.register_sub_command(CommandWalletImportWatchAddr())
-        self.register_sub_command(CommandWalletImportMultisigAddr())
-        self.register_sub_command(CommandWalletImportToken())
-        self.register_sub_command(CommandWalletImportContractAddr())
-
-    def command_desc(self):
-        return CommandDesc('import', 'import wallet items')
-
-    def execute(self, arguments):
-        item = PromptUtils.get_arg(arguments)
-
-        if not item:
-            print(f"Please specify an action. See help for available actions")
-            return False
-
-        try:
-            return self.execute_sub_command(item, arguments[1:])
-        except KeyError:
-            print(f"{item} is an invalid parameter")
-            return False
-
-
-class CommandWalletExportWIF(CommandBase):
-    def __init__(self):
-        super().__init__()
-
-    def execute(self, arguments):
-        wallet = PromptData.Wallet
-
-        if len(arguments) != 1:
-            print("Please specify the required parameter")
-            return False
-
-        address = arguments[0]
-        keys = wallet.GetKeys()
-        for key in keys:
-            if key.GetAddress() == address:
-                print(f"WIF: {key.Export()}")
-                return True
-        else:
-            print(f"Could not find address {address} in wallet")
-            return False
-
-    def command_desc(self):
-        p1 = ParameterDesc('address', 'public address in the wallet')
-        return CommandDesc('wif', 'export an unprotected private key record of an address', [p1])
-
-
-class CommandWalletExportNEP2(CommandBase):
-    def __init__(self):
-        super().__init__()
-
-    def execute(self, arguments):
-        wallet = PromptData.Wallet
-
-        if len(arguments) != 1:
-            print("Please specify the required parameters")
-            return False
-
-        address = arguments[0]
-
-        passphrase = prompt("[key password] ", is_password=True)
-        len_pass = len(passphrase)
-        if len_pass < 10:
-            print(f"Passphrase is too short, length: {len_pass}. Mininum length is 10")
-            return False
-
-        passphrase_confirm = prompt("[key password again] ", is_password=True)
-
-        if passphrase != passphrase_confirm:
-            print("Please provide matching passwords")
-            return False
-
-        keys = wallet.GetKeys()
-        for key in keys:
-            if key.GetAddress() == address:
-                print(f"NEP2: {key.ExportNEP2(passphrase)}")
-                return True
-        else:
-            print(f"Could not find address {address} in wallet")
-            return False
-
-    def command_desc(self):
-        p1 = ParameterDesc('address', 'public address in the wallet')
-        return CommandDesc('nep2', 'export a passphrase protected private key record of an address (NEP-2 format)', [p1])
-
-
-class CommandWalletImportWIF(CommandBase):
-    def __init__(self):
-        super().__init__()
-
-    def execute(self, arguments):
-        wallet = PromptData.Wallet
-
-        if len(arguments) != 1:
-            print("Please specify the required parameter")
-            return False
-
-        wif = arguments[0]
-        try:
-            kp = KeyPair.PrivateKeyFromWIF(wif)
-        except ValueError as e:
-            print(f"WIF Error: {str(e)}")
-            return False
-
-        try:
-            key = wallet.CreateKey(kp)
-            print(f"Imported key: {wif}")
-            pub_key = key.PublicKey.encode_point(True).decode('utf-8')
-            print(f"Pubkey: {pub_key}")
-            print(f"Address: {key.GetAddress()}")
-
-        except Exception as e:
-            # couldn't find an exact call that throws this but it was in the old code. Leaving it in for now.
-            print(f"Key creation error: {str(e)}")
-            return False
-
-        return True
-
-    def command_desc(self):
-        p1 = ParameterDesc('key', 'private key record in WIF format')
-        return CommandDesc('wif', 'import an unprotected private key record of an address', [p1])
-
-
-class CommandWalletImportNEP2(CommandBase):
-    def __init__(self):
-        super().__init__()
-
-    def execute(self, arguments):
-        wallet = PromptData.Wallet
-
-        if len(arguments) != 1:
-            print("Please specify the required parameters")
-            return False
-
-        nep2_key = arguments[0]
-        passphrase = prompt("[key password] ", is_password=True)
-
-        try:
-            kp = KeyPair.PrivateKeyFromNEP2(nep2_key, passphrase)
-        except ValueError as e:
-            print(str(e))
-            return False
-
-        try:
-            key = wallet.CreateKey(kp)
-            print(f"Imported key: {nep2_key}")
-            pub_key = key.PublicKey.encode_point(True).decode('utf-8')
-            print(f"Pubkey: {pub_key}")
-            print(f"Address: {key.GetAddress()}")
-
-        except Exception as e:
-            # couldn't find an exact call that throws this but it was in the old code. Leaving it in for now.
-            print(f"Key creation error: {str(e)}")
-            return False
-
-    def command_desc(self):
-        p1 = ParameterDesc('private key', 'a NEP-2 protected private key')
-        return CommandDesc('nep2', 'import a passphrase protected private key record (NEP-2 format)', [p1])
-
-
-class CommandWalletImportWatchAddr(CommandBase):
-    def __init__(self):
-        super().__init__()
-
-    def execute(self, arguments):
-        wallet = PromptData.Wallet
-
-        if len(arguments) != 1:
-            print("Please specify the required parameters")
-            return False
-
-        addr = arguments[0]
-        if not isValidPublicAddress(addr):
-            print("Invalid address specified")
-            return False
-
-        try:
-            addr_script_hash = wallet.ToScriptHash(addr)
-            wallet.AddWatchOnly(addr_script_hash)
-        except ValueError as e:
-            print(str(e))
-            return False
-
-        print(f"Added address {addr} as watch-only")
-        return True
-
-    def command_desc(self):
-        p1 = ParameterDesc('address', 'a public NEO address to watch')
-        return CommandDesc('watch_addr', 'import a public address as watch only', [p1])
-
-
-class CommandWalletImportMultisigAddr(CommandBase):
-    def __init__(self):
-        super().__init__()
-
-    def execute(self, arguments):
-        wallet = PromptData.Wallet
-
-        if len(arguments) < 3:
-            print("Please specify the minimum required parameters")
-            return False
-
-        pubkey_in_wallet = arguments[0]
-        if not _is_valid_public_key(pubkey_in_wallet):
-            print("Invalid public key format")
-            return False
-
-        key_script_hash = Crypto.ToScriptHash(pubkey_in_wallet, unhex=True)
-        if not wallet.ContainsKeyHash(key_script_hash):
-            print("Supplied first public key does not exist in own wallet.")
-            return False
-
-        try:
-            min_signature_cnt = int(arguments[1])
-        except ValueError:
-            print(f"Invalid minimum signature count value: {arguments[1]}")
-            return False
-
-        if min_signature_cnt < 1:
-            print("Minimum signatures count cannot be lower than 1")
-            return False
-
-        # validate minimum required signing key count
-        signing_keys = arguments[2:]
-        len_signing_keys = len(signing_keys)
-        if len_signing_keys < min_signature_cnt:
-            # we need at least 2 public keys in total otherwise it's just a regular address.
-            # 1 pub key is from an address in our own wallet, a secondary key can come from any place.
-            print(f"Missing remaining signing keys. Minimum required: {min_signature_cnt} given: {len_signing_keys}")
-            return False
-
-        # validate remaining pub keys
-        for key in signing_keys:
-            if not _is_valid_public_key(key):
-                print(f"Invalid signing key {key}")
-                return False
-
-        signing_keys.append(pubkey_in_wallet)
-
-        # validate that all signing keys are unique
-        if len(signing_keys) > len(set(signing_keys)):
-            print("Provided signing keys are not unique")
-            return False
-
-        verification_contract = Contract.CreateMultiSigContract(key_script_hash, min_signature_cnt, signing_keys)
-        address = verification_contract.Address
-        wallet.AddContract(verification_contract)
-        print(f"Added multi-sig contract address {address} to wallet")
-        return True
-
-    def command_desc(self):
-        p1 = ParameterDesc('own pub key', 'public key in your own wallet (use `wallet` to find the information)')
-        p2 = ParameterDesc('sign_cnt', 'the minimum number of signatures required for using the address (min is: 1)')
-        p3 = ParameterDesc('signing key n', 'all remaining signing public keys')
-        return CommandDesc('multisig_addr', 'import a multi-signature address', [p1, p2, p3])
-
-
-class CommandWalletImportToken(CommandBase):
-    def __init__(self):
-        super().__init__()
-
-    def execute(self, arguments):
-        if len(arguments) != 1:
-            print("Please specify the required parameter")
-            return
-
-        try:
-            contract_hash = UInt160.ParseString(arguments[0]).ToBytes()
-        except Exception:
-            print(f"Invalid contract hash: {arguments[0]}")
-            return
-
-        return ImportToken(PromptData.Wallet, contract_hash)
-
-    def command_desc(self):
-        p1 = ParameterDesc('contract_hash', 'the token contract hash')
-        return CommandDesc('token', 'import a token', [p1])
-
-
-class CommandWalletImportContractAddr(CommandBase):
-    def __init__(self):
-        super().__init__()
-
-    def execute(self, arguments):
-        wallet = PromptData.Wallet
-
-        if len(arguments) != 2:
-            print("Please specify the required parameters")
-            return
-
-        try:
-            contract_hash = UInt160.ParseString(arguments[0]).ToBytes()
-        except Exception:
-            print(f"Invalid contract hash: {arguments[0]}")
-            return
-
-        pubkey = arguments[1]
-        if not _is_valid_public_key(pubkey):
-            print(f"Invalid pubkey: {arguments[1]}")
-            return
-
-        pubkey_script_hash = Crypto.ToScriptHash(pubkey, unhex=True)
-
-        return ImportContractAddr(wallet, contract_hash, pubkey_script_hash)
-
-    def command_desc(self):
-        p1 = ParameterDesc('contract_hash', 'hash of the contract')
-        p2 = ParameterDesc('pubkey', 'pubkey of the contract')
-        return CommandDesc('contract_addr', 'import a contract address', [p1, p2])
-
-
 #########################################################################
 #########################################################################
-
-
-def CreateAddress(wallet, args):
-    try:
-        int_args = int(args)
-    except (ValueError, TypeError) as error:  # for non integer args or Nonetype
-        print(error)
-        return
-
-    if wallet is None:
-        print("Please open a wallet.")
-        return
-
-    if int_args <= 0:
-        print('Enter a number greater than 0.')
-        return
-
-    address_list = []
-    for _ in range(int_args):
-        keys = wallet.CreateKey()
-        account = Account.get(PublicKeyHash=keys.PublicKeyHash.ToBytes())
-        address_list.append(account.contract_set[0].Address.ToString())
-    print("Created %s new addresses: " % int_args, address_list)
-    return wallet
-
-
-def DeleteAddress(wallet, addr):
-    try:
-        scripthash = wallet.ToScriptHash(addr)
-        error_str = ""
-
-        success, _ = wallet.DeleteAddress(scripthash)
-    except ValueError as e:
-        success = False
-        error_str = f" with error: {e}"
-
-    if success:
-        print(f"Deleted address {addr}")
-    else:
-        print(f"Error deleting addr {addr}{error_str}")
-
-    return success
-
-
-def ImportToken(wallet, contract_hash):
-    if wallet is None:
-        print("please open a wallet")
-        return
-
-    contract = Blockchain.Default().GetContract(contract_hash)
-
-    if contract:
-        hex_script = binascii.hexlify(contract.Code.Script)
-        token = NEP5Token.NEP5Token(script=hex_script)
-
-        result = token.Query()
-
-        if result:
-            wallet.AddNEP5Token(token)
-            print("added token %s " % json.dumps(token.ToJson(), indent=4))
-            return token
-        else:
-            print("Could not import token")
-    else:
-        print("Could not find the contract hash")
-
-
-def AddAlias(wallet, addr, title):
-    if wallet is None:
-        print("Please open a wallet")
-        return False
-
-    try:
-        script_hash = wallet.ToScriptHash(addr)
-        wallet.AddNamedAddress(script_hash, title)
-        return True
-    except Exception as e:
-        print(e)
-        return False
 
 
 def ClaimGas(wallet, require_password=True, from_addr_str=None):
@@ -954,92 +390,3 @@ def ShowUnspentCoins(wallet, asset_id=None, from_addr=None, watch_only=False, do
         print("No unspent assets matching the arguments.")
 
     return unspents
-
-
-def SplitUnspentCoin(wallet, asset_id, from_addr, index, divisions, fee=Fixed8.Zero(), prompt_passwd=True):
-    """
-    Split unspent asset vins into several vouts
-
-    Args:
-        wallet (neo.Wallet): wallet to show unspent coins from.
-        asset_id (UInt256): a bytearray (len 32) representing an asset on the blockchain.
-        from_addr (UInt160): a bytearray (len 20) representing an address.
-        index (int): index of the unspent vin to split
-        divisions (int): number of vouts to create
-        fee (Fixed8): A fee to be attached to the Transaction for network processing purposes.
-        prompt_passwd (bool): prompt password before processing the transaction
-
-    Returns:
-        neo.Core.TX.Transaction.ContractTransaction: contract transaction created
-    """
-
-    if wallet is None:
-        print("Please open a wallet.")
-        return
-
-    unspent_items = wallet.FindUnspentCoinsByAsset(asset_id, from_addr=from_addr)
-    if not unspent_items:
-        print(f"No unspent assets matching the arguments.")
-        return
-
-    if index < len(unspent_items):
-        unspent_item = unspent_items[index]
-    else:
-        print(f"Could not find unspent item for asset {asset_id} with index {index}")
-        return
-
-    outputs = split_to_vouts(asset_id, from_addr, unspent_item.Output.Value, divisions)
-
-    # subtract a fee from the first vout
-    if outputs[0].Value > fee:
-        outputs[0].Value -= fee
-    else:
-        print("Fee could not be subtracted from outputs.")
-        return
-
-    contract_tx = ContractTransaction(outputs=outputs, inputs=[unspent_item.Reference])
-
-    ctx = ContractParametersContext(contract_tx)
-    wallet.Sign(ctx)
-
-    print("Splitting: %s " % json.dumps(contract_tx.ToJson(), indent=4))
-    if prompt_passwd:
-        passwd = prompt("[Password]> ", is_password=True)
-        if not wallet.ValidatePassword(passwd):
-            print("incorrect password")
-            return
-
-    if ctx.Completed:
-        contract_tx.scripts = ctx.GetScripts()
-
-        relayed = NodeLeader.Instance().Relay(contract_tx)
-
-        if relayed:
-            wallet.SaveTransaction(contract_tx)
-            print("Relayed Tx: %s " % contract_tx.Hash.ToString())
-            return contract_tx
-        else:
-            print("Could not relay tx %s " % contract_tx.Hash.ToString())
-
-
-def split_to_vouts(asset, addr, input_val, divisions):
-    divisor = Fixed8(divisions)
-
-    new_amounts = input_val / divisor
-    outputs = []
-    total = Fixed8.Zero()
-
-    if asset == Blockchain.Default().SystemShare().Hash:
-        if new_amounts % Fixed8.FD() > Fixed8.Zero():
-            new_amounts = new_amounts.Ceil()
-
-    while total < input_val:
-        if total + new_amounts < input_val:
-            outputs.append(TransactionOutput(asset, new_amounts, addr))
-            total += new_amounts
-        else:
-            diff = input_val - total
-            outputs.append(TransactionOutput(asset, diff, addr))
-            total += diff
-
-    return outputs
